Skiller ut tilbakestill behandling til behandlingsresultat i egen service for å hindre sykel i avhengigheter
