+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/eøs/endringsabonnement/TilbakestillBehandling.kt
@@ -5,32 +5,35 @@ import no.nav.familie.ba.sak.kjerne.eøs.felles.PeriodeOgBarnSkjemaEndringAbonne
 import no.nav.familie.ba.sak.kjerne.eøs.kompetanse.domene.Kompetanse
 import no.nav.familie.ba.sak.kjerne.eøs.utenlandskperiodebeløp.UtenlandskPeriodebeløp
 import no.nav.familie.ba.sak.kjerne.eøs.valutakurs.Valutakurs
-import no.nav.familie.ba.sak.kjerne.steg.TilbakestillBehandlingService
+import no.nav.familie.ba.sak.kjerne.steg.TilbakestillBehandlingTilBehandlingsresultatService
 import org.springframework.stereotype.Service
 
 @Service
 class TilbakestillBehandlingFraKompetanseEndringService(
-    private val tilbakestillBehandlingService: TilbakestillBehandlingService
+    private val tilbakestillBehandlingTilBehandlingsresultatService: TilbakestillBehandlingTilBehandlingsresultatService
 ) : PeriodeOgBarnSkjemaEndringAbonnent<Kompetanse> {
     override fun skjemaerEndret(behandlingId: BehandlingId, endretTil: Collection<Kompetanse>) {
-        tilbakestillBehandlingService.tilbakestillBehandlingTilBehandlingsresultat(behandlingId.id)
+        tilbakestillBehandlingTilBehandlingsresultatService
+            .tilbakestillBehandlingTilBehandlingsresultat(behandlingId.id)
     }
 }
 
 @Service
 class TilbakestillBehandlingFraUtenlandskPeriodebeløpEndringService(
-    private val tilbakestillBehandlingService: TilbakestillBehandlingService
+    private val tilbakestillBehandlingTilBehandlingsresultatService: TilbakestillBehandlingTilBehandlingsresultatService
 ) : PeriodeOgBarnSkjemaEndringAbonnent<UtenlandskPeriodebeløp> {
     override fun skjemaerEndret(behandlingId: BehandlingId, endretTil: Collection<UtenlandskPeriodebeløp>) {
-        tilbakestillBehandlingService.tilbakestillBehandlingTilBehandlingsresultat(behandlingId.id)
+        tilbakestillBehandlingTilBehandlingsresultatService
+            .tilbakestillBehandlingTilBehandlingsresultat(behandlingId.id)
     }
 }
 
 @Service
 class TilbakestillBehandlingFraValutakursEndringService(
-    private val tilbakestillBehandlingService: TilbakestillBehandlingService
+    private val tilbakestillBehandlingTilBehandlingsresultatService: TilbakestillBehandlingTilBehandlingsresultatService
 ) : PeriodeOgBarnSkjemaEndringAbonnent<Valutakurs> {
     override fun skjemaerEndret(behandlingId: BehandlingId, endretTil: Collection<Valutakurs>) {
-        tilbakestillBehandlingService.tilbakestillBehandlingTilBehandlingsresultat(behandlingId.id)
+        tilbakestillBehandlingTilBehandlingsresultatService
+            .tilbakestillBehandlingTilBehandlingsresultat(behandlingId.id)
     }
 }
