+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/steg/TilbakestillBehandlingService.kt
@@ -98,33 +98,4 @@ class TilbakestillBehandlingService(
             steg = StegType.VILKÅRSVURDERING
         )
     }
-
-    /**
-     * Når en andel vurderes (endres) vil vi resette steget og slette data som blir generert senere i løypa
-     */
-    @Transactional
-    fun tilbakestillBehandlingTilBehandlingsresultat(behandlingId: Long): Behandling {
-        val behandling = behandlingHentOgPersisterService.hent(behandlingId)
-
-        if (behandling.erTilbakestiltTilBehandlingsresultat()) {
-            return behandling
-        }
-
-        vedtaksperiodeHentOgPersisterService.slettVedtaksperioderFor(
-            vedtak = vedtakRepository.findByBehandlingAndAktiv(
-                behandlingId
-            )
-        )
-        tilbakekrevingService.slettTilbakekrevingPåBehandling(behandlingId)
-        return behandlingService.leggTilStegPåBehandlingOgSettTidligereStegSomUtført(
-            behandlingId = behandlingId,
-            steg = StegType.BEHANDLINGSRESULTAT
-        )
-    }
-}
-
-fun Behandling.erTilbakestiltTilBehandlingsresultat(): Boolean {
-    val gjeldendeSteg = this.behandlingStegTilstand.last()
-    return gjeldendeSteg.behandlingSteg == StegType.BEHANDLINGSRESULTAT &&
-        gjeldendeSteg.behandlingStegStatus == BehandlingStegStatus.IKKE_UTFØRT
 }
